backend/requirements.txt
# --- Core Django ---
Django>=5.0,<6.0

# --- Database ---
psycopg[binary]  # PostgreSQL driver

# --- WSGI server for production ---
gunicorn

# --- Environment variables ---
python-dotenv

# --- Django REST Framework (API) ---
djangorestframework
django-filter       # filters for DRF
djangorestframework-simplejwt  # JWT auth
drf-nested-routers # nested routers

# --- CORS (for React frontend API calls) ---
django-cors-headers

# --- Media/File support ---
Pillow
